Added role checks before data retrieval
